Refactor : command -> condition으로 패키지명 변경
